requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
IMDbPY==2020.9.25
joblib==1.0.1
lightfm==1.16
numpy==1.19.5
requests==2.25.1
scikit-learn==0.24.1
scipy==1.5.4
threadpoolctl==2.1.0
urllib3==1.26.4
